requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==4.0
django-crispy-forms==1.13.0
django-filter==21.1
django-ordered-model==3.5
idna==3.3
numpy==1.21.4
Pillow==9.0.0
pycodestyle==2.8.0
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.7
